Change Month type to int

+++ TruckProblem/Main.java
@@ -6,10 +6,10 @@ public class Main {
         Scanner scan = new Scanner(System.in);
         
         System.out.print("Enter Speed : ");
-        int speed = scan.nextInt();
+        double speed = scan.nextDouble();
         
         System.out.print("Enter Distance : ");
-        int distance = scan.nextInt();
+        double distance = scan.nextDouble();
         
         truck.setSpeed(speed);
         truck.setDistance(distance);
